Lista-5/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.122
botocore==1.34.122
Django==4.2.13
django-bootstrap-v5==1.0.11
django-environ==0.11.2
djangorestframework==3.15.1
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
